+++ src/roq/deribit/web_socket.hpp
@@ -51,6 +51,7 @@ struct WebSocket final : public web::socket::Client::Handler, public core::jsonr
   struct Handler {
     virtual void operator()(Trace<StreamStatus> const &) = 0;
     virtual void operator()(Trace<ExternalLatency> const &) = 0;
+    virtual void operator()(Trace<ReferenceData> const &, bool is_last) = 0;
     virtual void operator()(Trace<TopOfBook> const &, bool is_last) = 0;
     virtual void operator()(Trace<MarketStatus> const &, bool is_last) = 0;
     // cross-communication